requirements.txt
Flask==2.0.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Werkzeug==2.0.2
pyjwt==2.1.0
pymysql==1.0.2
SQLAlchemy==1.4.35
redis==3.5.3
Flask-Redis==0.4.0
